pyproject.toml
[project]
name = "albo-arco"
version = "0.1.0"
description = "Scrapes Albo Pretorio items and posts to Telegram"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "python-telegram-bot>=21.0",
    "python-dotenv>=1.0.0",
]
